requirements.txt
streamlit>=1.34.0
qdrant-client>=1.7.4
sentence-transformers>=2.7.0
torch>=2.2.0
transformers>=4.41.0
numpy>=1.26.0
pandas>=2.2.0
tqdm>=4.66.0
python-dotenv>=1.0.1
pypdf>=4.2.0
ollama>=0.1.8
zenml>=0.64.0

